Testing with Oracle db backend Had to convert all ints to longs All INTEGERs to BIGINTS
PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: http://svn.apache.org/repos/asf/portals/jetspeed-2/trunk@186388 13f79535-47bb-0310-9956-ffa450edef68

+++ commons/src/java/org/apache/jetspeed/om/common/impl/LocalizedFieldImpl.java
@@ -36,7 +36,7 @@ public class LocalizedFieldImpl implements LocalizedField
     protected Locale locale;
     
     protected long parentId;
-    protected int id;
+    protected long id;
     
 
     public LocalizedFieldImpl()
@@ -95,7 +95,7 @@ public class LocalizedFieldImpl implements LocalizedField
      */
     public void setId(String oid)
     {
-        id = JetspeedObjectID.createFromString(oid).intValue();
+        id = JetspeedObjectID.createFromString(oid).longValue();
     }
     
     public void setLanguage(String lang)
